+++ tools/run-test-build.sh
@@ -4,4 +4,4 @@ SCRIPT=`realpath "${0}"`
 SCRIPTPATH=`dirname "${SCRIPT}"`
 PROJECTROOT=`realpath "${SCRIPTPATH}"/..`
 
-cargo +nightly run --manifest-path "${PROJECTROOT}"/tools/test-build/Cargo.toml --release -- --repo "${PROJECTROOT}"
+cargo run --manifest-path "${PROJECTROOT}"/tools/test-build/Cargo.toml --release -- --repo "${PROJECTROOT}"